+++ gazebo/rendering/Scene.cc
@@ -652,8 +652,7 @@ UserCameraPtr Scene::GetUserCamera(uint32_t index) const
 //////////////////////////////////////////////////
 void Scene::RemoveCamera(const std::string &_name)
 {
-  std::vector<CameraPtr>::iterator iter;
-  for (iter = this->dataPtr->cameras.begin();
+  for (auto iter = this->dataPtr->cameras.begin();
       iter != this->dataPtr->cameras.end(); ++iter)
   {
     if ((*iter)->GetName() == _name)