+++ gazebo/sensors/CameraSensor.cc
@@ -113,8 +113,7 @@ void CameraSensor::Init()
       }
     }
 
-    this->camera = this->scene->CreateCamera(
-        this->sdf->Get<std::string>("name"), false);
+    this->camera = this->scene->CreateCamera(this->GetScopedName(), false);
 
     if (!this->camera)
     {
@@ -134,7 +133,7 @@ void CameraSensor::Init()
     }
 
     this->camera->Init();
-    this->camera->CreateRenderTexture(this->GetName() + "_RttTex");
+    this->camera->CreateRenderTexture(this->GetScopedName() + "_RttTex");
     ignition::math::Pose3d cameraPose = this->pose;
     if (cameraSdf->HasElement("pose"))
       cameraPose = cameraSdf->Get<ignition::math::Pose3d>("pose") + cameraPose;
@@ -170,7 +169,7 @@ void CameraSensor::Fini()
 
   if (this->camera)
   {
-    this->scene->RemoveCamera(this->camera->GetName());
+    this->scene->RemoveCamera(this->GetScopedName());
   }
 
   this->camera.reset();