+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Gazebo 6.X.X (201X-XX-XX)
 
+1. Fix removing multiple camera sensors that have the same camera name
+    * [Pull request #](https://bitbucket.org/osrf/gazebo/pull-request/)
+
 1. Backport model editor toolbar fixed joint option from [pull request #1794](https://bitbucket.org/osrf/gazebo/pull-request/1794)
     * [Pull request #1957](https://bitbucket.org/osrf/gazebo/pull-request/1957)